+++ app/datatables/components_datatable.rb
@@ -3,7 +3,7 @@ class ComponentsDatatable < IndexDatatable
     super
 
     columns([:score_name, :competition_name, :competition_class, :competition_type,
-             :category, :segment, :season,
+             :category, :segment, :date, :season,
              :ranking, :skater_name, :nation,
              :number, :name, :factor, :judges, :value,])
 
@@ -16,6 +16,7 @@ class ComponentsDatatable < IndexDatatable
       category: "scores.category",
       segment: "scores.segment",
       #date: "scores.performed_segment.starting_time",
+      date: "scores.date",
       ranking: "scores.ranking",
       skater_name: "skaters.name",
       nation: "skaters.nation",
@@ -23,7 +24,7 @@ class ComponentsDatatable < IndexDatatable
     }
   
     ## searchble
-    #columns[:date].searchable = false
+    columns[:date].searchable = false
     columns[:value].operator = params[:value_operator].presence || :eq      if view_context
 
     ## visible