+++ app/datatables/elements_datatable.rb
@@ -3,7 +3,7 @@ class ElementsDatatable < IndexDatatable
     super
 
     columns([:score_name, :competition_name, :competition_class, :competition_type,
-             #:category, :segment, :date, :season,
+             :category, :segment, :date, :season,
              :category, :segment, :season,
              :skater_name, :nation,
              :number, :name, :element_type, :level, :credit, :info, :base_value, :goe, :judges, :value,])
@@ -17,13 +17,14 @@ class ElementsDatatable < IndexDatatable
       category: "scores.category",
       segment: "scores.segment",
       #date: "scores.performed_starting_time.starting_time",
+      date: "scores.date",
       skater_name: "skaters.name",
       nation: "skaters.nation",
       name: "elements.name",
       base_value: "elements.base_value",
     }
     ## searchable
-    [:credit, :info].each {|key| columns[key].searchable = false }    
+    [:date, :credit, :info].each {|key| columns[key].searchable = false }    
 
     ## visible
     [:competition_class, :competition_type].each {|key|