+++ app/decorators/score_decorator.rb
@@ -34,10 +34,12 @@ class ScoreDecorator < EntryDecorator
     h.link_to(model.name, controller: :scorecalc, score_name: model.name)
   end
 =end
+=begin
   def segment_starting_time
     #l(model.performed_segment.starting_time.in_time_zone(competition.timezone))
     l(model.segment_starting_time)
   end
+=end
 =begin
   def youtube_search
     h.link_to("Youtube", "http://www.youtube.com/results?q=" + [score.skater.name, score.competition.name, score.segment].join('+'), target: "_blank")