+++ app/updaters/competition_updater.rb
@@ -41,15 +41,13 @@ class CompetitionUpdater
 
           parsed[:segments][category].each do |segment, seg_item|
             ###
-            competition.performed_segments.create do |ps|
+            ps = competition.performed_segments.create do |ps|
               ps.category = category
               ps.segment = segment
               ps.starting_time = seg_item[:time]
-
-            ###
-            #update_score(competition, category, segment, seg_item[:score_url], seg_item[:result_url], segment_starting_time: seg_item[:time])
-              update_score(competition, category, segment, seg_item[:score_url], seg_item[:result_url], {performed_segment: ps})
             end
+            ###
+            update_score(competition, category, segment, seg_item[:score_url], seg_item[:result_url], date: ps.starting_time.to_date)
           end
         end
       end
@@ -78,7 +76,7 @@ class CompetitionUpdater
     end
   end 
   ################
-  def update_score(competition, category, segment, score_url, result_url, extr_params = {})
+  def update_score(competition, category, segment, score_url, result_url, date:)
     segment_results = @parser.parse_segment_result(result_url)
 
     @parser.parse_score(score_url).each do |sc_parsed|
@@ -99,8 +97,9 @@ class CompetitionUpdater
           score.attributes = {
             category_result: relevant_cr,
             skater: skater,
+            date: date,
             #segment_starting_time: segment_starting_time,
-          }.merge(extr_params)
+          }
           attrs = score.class.column_names.map(&:to_sym) & sc_parsed.keys
           score.attributes = sc_parsed.slice(*attrs)
 
@@ -112,31 +111,6 @@ class CompetitionUpdater
           score.update(elements_summary: score.elements.map(&:name).join('/'))
           score.update(components_summary: score.components.map(&:value).join('/'))
           puts score.summary if @verbose
-          
-=begin
-              ## judge details
-              if self.start_date > Time.zone.parse("2016-7-1") # was random order in the past
-                ### elements
-                score.elements.each do |element|
-                  element.judges.split(/\s/).each_with_index do |value, i|
-                    #next if panels[:judges].count <= i+1
-                    element.element_judge_details.create(panel_name: panels[:judges][i+1][:name],
-                                                         panel_nation: panels[:judges][i+1][:nation],
-                                                         number: i, value: value)
-                  end
-                end
-
-                ### component
-                score.components.each do |component|
-                  component.judges.split(/\s/).each_with_index do |value, i|
-                    #next if panels[:judges].count <= i+1
-                    component.component_judge_details.create(panel_name: panels[:judges][i+1][:name],
-                                                             panel_nation: panels[:judges][i+1][:nation],
-                                                             number: i, value: value)
-                  end
-                end
-              end
-=end
         }
       end
     end